Add update on every run, runtime info

+++ lib/refresh.dart
@@ -1,5 +1,3 @@
-import 'dart:developer';
-
 import 'package:alfred_noteplan_fts_refresh/config.dart';
 import 'package:alfred_noteplan_fts_refresh/db_cache.dart';
 import 'package:alfred_noteplan_fts_refresh/db_fts.dart';
@@ -8,23 +6,24 @@ import 'package:sqlite3/sqlite3.dart';
 
 import 'package:alfred_noteplan_fts_refresh/note.dart';
 
-void refresh({bool force = false}) {
-	// print('Using sqlite3 ${sqlite3.version.libVersion}');
+int refresh(DbFts db, {bool force = false}) {
 	final cache = DbCache(sqlite3.open(Config.path_cache_db));
-	final db    = DbFts(sqlite3.open('database.sqlite3'));
-
-	// Ensure existing notes database
-	db.ensure_setup();
 
 	// Get changed notes in Cache
 	initializeDateFormatting('en_GB', null);
 	List<Note> new_notes = [];
-	for (var result in cache.select_updated()) { new_notes.add(Note(result)); }
+	for (var result in cache.select_updated(since: force ? 0 : db.get_last_update())) { new_notes.add(Note(result)); }
+
+	// Delete/reinsert the notes
+	if (new_notes.isNotEmpty) {
+		db.delete_notes(new_notes.map((e) => e.filename));
+		db.insert_notes(new_notes);
+	}
+
+	// Save the last update and cleanup
+	db.set_last_update();
 	cache.dispose();
 
-	// Delete existing versions of updated notes
-	db.delete_notes(new_notes.map((e) => e.filename));
-	db.insert_notes(new_notes);
-
-	db.dispose();
+	// return number of updates
+	return new_notes.length;
 }
